Improvements on the generation of silica detritus (#159)
Adding flexibility to the routing of phytoplankton silicon losses to enable increases in the production of Si detritus.

First, we implemented a Si loss associated with negative growth, to be consistent with how N, P, and Fe losses due to respiration exceeding growth is handled.

Second, we added a scaling factor (phi_sidiss_mort) that determines the amount of silica test left over in medium and large phytoplankton as phytoplankton dies. To target an overall dissolution rate of 0.01 day-1 at 0 degrees C (consistent with detritus) we can set this value to be 0.67. However, to reproduce answers, here in this PR we have set phi_sidiss_mort to be 0.

Lastly, we added flexibility to what percentage of Si ingestion by zooplankton is egested to detritus. By default, the values of phi_det_si for small, medium, and large zooplankton are the same as phi_det, but this parameter can also be increased to route more silicon ingestion to detritus. For ESM4.5 we are testing values 5-10% more than the default phi_det values.